complete dalloc syscall and there exists problem in outputing kernal memory addr which is not permitted by sys

+++ kernel/sysproc.c
@@ -10,7 +10,7 @@ uint64
 sys_dalloc(void){
   int size;
   argint(0,&size);
-  return dalloc(size);
+  return (uint64)dalloc(size);
 }
 
 uint64
